backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose==3.3.0
PyJWT==2.9.0
argon2-cffi==23.1.0
cryptography==43.0.1
python-multipart==0.0.9
pydantic-settings==2.5.2
email-validator==2.2.0
pytest==8.3.3
httpx==0.27.2
